[CRUD] creacion
