+++ frontend/src/routes/CRUD.jsx
@@ -9,20 +9,10 @@ export const CRUD = () => {
   const links = [];
   const components = [];
 
-  // Opciones de cada columna
-  const columnOptions = ['', '', '', '', '', ''];
-
   return (
     <div>
       <CustomNavbar links={links} components={components}/>
-      <ResultTable 
-        fil1={'X'}
-        fil2={'X'}
-        fil3={'X'}
-        fil4={'X'}
-        fil5={'X'}
-        order={'autor'}
-        hier={'ASC'}/>
+      <ResultTable/>
     </div>
   );
 }
